+++ src/game.h
@@ -450,8 +450,8 @@ public:
 	void addAnimatedText(std::string_view message, const Position& pos, TextColor_t color);
 	static void addAnimatedText(const SpectatorVec& list, std::string_view message, const Position& pos,
 	                            TextColor_t color);
-	void addMagicEffect(const Position& pos, uint8_t effect);
-	static void addMagicEffect(const SpectatorVec& spectators, const Position& pos, uint8_t effect);
+	void addMagicEffect(const Position& pos, uint16_t effect);
+	static void addMagicEffect(const SpectatorVec& spectators, const Position& pos, uint16_t effect);
 	void addDistanceEffect(const Position& fromPos, const Position& toPos, uint8_t effect);
 	static void addDistanceEffect(const SpectatorVec& spectators, const Position& fromPos, const Position& toPos,
 	                              uint8_t effect);
