+++ src/player.h
@@ -798,7 +798,7 @@ public:
 			client->sendIcons(getClientIcons());
 		}
 	}
-	void sendMagicEffect(const Position& pos, uint8_t type) const
+	void sendMagicEffect(const Position& pos, uint16_t type) const
 	{
 		if (client) {
 			client->sendMagicEffect(pos, type);
