+++ src/protocolgame.h
@@ -149,7 +149,7 @@ private:
 	void sendFYIBox(std::string_view message);
 
 	void sendDistanceShoot(const Position& from, const Position& to, uint8_t type);
-	void sendMagicEffect(const Position& pos, uint8_t type);
+	void sendMagicEffect(const Position& pos, uint16_t type);
 	void sendCreatureHealth(const Creature* creature);
 	void sendSkills();
 	void sendPing();
